Restructure project (component oriented), Add animation to Views
