+++ js/Components/AppList/App.js
@@ -2,7 +2,7 @@
  * Created by tema on 19.11.16.
  */
 
-const Artemone = require('./../Artemone/Artemone');
+const Artemone = require('./../../Artemone/Artemone');
 
 class App extends Artemone.Models {
 	constructor(attributes) {
