+++ js/Components/Comment/Comment.js
@@ -2,7 +2,7 @@
  * Created by tema on 11.11.16.
  */
 
-const Artemone = require('./../Artemone/Artemone');
+const Artemone = require('./../../Artemone/Artemone');
 
 class Comment extends Artemone.Models {
 	constructor(attributes) {
