+++ js/app.js
@@ -1,35 +1,25 @@
 (function () {
 	const Artemone = require('./Artemone/Artemone');
-	const AppView = require('./Views/App-view');
-	const AppModel = require('./Models/App');
-	const Lists = require('./Collections/Lists');
+	const AppView = require('./Components/AppList/AppView');
+	const AppModel = require('./Components/AppList/App');
+	const Lists = require('./Components/AppList/AppCollection');
 	const router = new Artemone.Router();
 
-	// const appModel = new AppModel({ title: 'aTodos' });
-
+	// Главное View приложения
 	const appView = new AppView();
 	const lists = new Lists();
 	lists.setName('lists');
 	appView.attach('.lists').setModel(lists);
-
-	appView.model.load().then((result) => {
-			result.loadTodos();
-		},
-		(error) => {
-			const appModel = new AppModel({ title: 'Todos' });
-			appView.model.add(appModel);
-			console.log(error);
-		});
+	appView.model.load();
 
 
-	const add_list = document.querySelector('.add-list');
-
+	// Добавление новых вью
 	function addList() {
 		const newAppModel = new AppModel({ title: 'Todos' });
 		appView.model.add(newAppModel);
 		appView.model.loadTodo(newAppModel);
 	}
-
+	const add_list = document.querySelector('.add-list');
 	add_list.addEventListener('click', addList);
 
 	router
